requirements.txt
google-genai>=0.8.0
Pillow>=10.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.9
mangum>=0.17.0